Abstract the commandbar from the modal itself so the bar can be used elsewhere
